course detail

+++ Online_Academy/Models/History.cs
@@ -17,7 +17,7 @@ namespace Online_Academy.Models
         public int id_course { get; set; }
         public int id_user { get; set; }
         public Nullable<System.DateTime> date { get; set; }
-        public Nullable<int> price { get; set; }
+        public Nullable<double> price { get; set; }
     
         public virtual Course Course { get; set; }
         public virtual User User { get; set; }
